Revert "[update]create image.go"
